% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignFeature.R
\name{alignFeature}
\alias{alignFeature}
\title{Title: align the features across samples}
\usage{
alignFeature(
  file_dir,
  filePattern,
  align_mz_tol = 0.01,
  align_rt_tol = 20,
  align_num = FALSE
)
}
\arguments{
\item{file_dir}{string, the folder storing the feature table to align}

\item{filePattern}{string, common file names of the feature tables}

\item{align_mz_tol}{numeric, m/z tolerance in Dalton}

\item{align_rt_tol}{numeric, rt tolerance in seconds}

\item{align_num}{logic, true or false when align the features}
}
\value{
data frame
}
\description{
Title: align the features across samples
}
