% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateFeature.R
\name{annonateFeature}
\alias{annonateFeature}
\title{Title}
\usage{
annonateFeature(
  featureTable,
  Cl_db,
  ion_mode = "P",
  ref_mz_tol = 25,
  dp_score = 70,
  dp_num = 2
)
}
\arguments{
\item{featureTable}{data frame, peaks}

\item{Cl_db}{database, check "db_demo"}

\item{ion_mode}{string, ionization mode "P" for positive mode}

\item{ref_mz_tol}{numeric, mass tolerance in Da}

\item{dp_score}{MS2 spectra similarity score threshold}

\item{dp_num}{number of matched fragments}
}
\value{
data frame
}
\description{
Title
}
