% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunction.R
\name{createMS1folder}
\alias{createMS1folder}
\title{Title: create folder to save EIC and MS1 spectra}
\usage{
createMS1folder(mzMLdirectory, mzMLfile)
}
\arguments{
\item{mzMLdirectory}{string, file directory}

\item{mzMLfile}{string, file full name .mzML}
}
\value{
string, folder name
}
\description{
Title: create folder to save EIC and MS1 spectra
}
