% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNetwork.R
\name{createNetwork}
\alias{createNetwork}
\title{Title: establish network}
\usage{
createNetwork(
  featureTable,
  seed_tb,
  reaction_pathways,
  reaction_pathway_mz_tol = 0.01,
  nw_spectra_score = 0.5,
  nw_spectra_match_num = 3
)
}
\arguments{
\item{featureTable}{data frame, all peaks}

\item{seed_tb}{data frame, reaction precursors or confidently annotated features}

\item{reaction_pathways}{data frame, possible reaction pathways}

\item{reaction_pathway_mz_tol}{numeric, m/z tolerance in reaction pathway}

\item{nw_spectra_score}{numeric, GNPS spectra similarity threshold}

\item{nw_spectra_match_num}{numeric, least number of matched fragments}
}
\value{
list, all network and explainable network
}
\description{
Title: establish network
}
