% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_demo.R
\docType{data}
\name{db_demo}
\alias{db_demo}
\title{Demon of database to match the tandem ms2 spectra}
\format{
A data frame with 1 row and 8 variables
\describe{
     \item{Name}{chemical names}
     \item{Precursor_type}{adduct form of precursors}
     \item{Collision_energy}{collision energy for fragmentation }
     \item{Precursor_mz}{mz value of the precursors}
     \item{Formula}{molecular formula}
     \item{Ion_mode}{ion mode}
     \item{MS2mz}{ms2 fragment mz}
     \item{MS2int}{ms2 fragment intensity}
     }
}
\source{
{Created in-house to serve as an example}
}
\usage{
db_demo
}
\description{
Contains 8 qualities of the chemicals
}
\examples{
data(db_demo) # Lazy loading. Data becomes visible as soon as
}
\keyword{datasets}
