% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNetwork.R
\name{dotProduct}
\alias{dotProduct}
\title{Title: dot product function
exp_fragMz, exp_fragInt, ref_fragMz, ref_fragMzInt: separate by ";"}
\usage{
dotProduct(exp_fragMz, exp_fragInt, ref_fragMz, ref_fragMzInt, ms2_tol = 0.02)
}
\arguments{
\item{exp_fragMz}{string, experimental fragment m/z separated by ";"}

\item{exp_fragInt}{string, experimental fragment intensity separated by ";"}

\item{ref_fragMz}{string,  reference fragment m/z separated by ";"}

\item{ref_fragMzInt}{string,  reference fragment intensity separated by ";"}

\item{ms2_tol}{numeric, m/z tolerance in Da}
}
\value{
list, dp score, count of matched fragment, mz
}
\description{
Title: dot product function
exp_fragMz, exp_fragInt, ref_fragMz, ref_fragMzInt: separate by ";"
}
