% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractEIC.R
\name{EIC_matrix}
\alias{EIC_matrix}
\title{Title: extract EIC}
\usage{
EIC_matrix(rawlcms, mz, rt, mz_tol = 0.01, rt_tol = 30)
}
\arguments{
\item{rawlcms}{xcms object processed by xcmsRaw}

\item{mz}{numeric, mass}

\item{rt}{numeric, retention time in minutes}

\item{mz_tol}{numeric, mass tolerance to extract EIC}

\item{rt_tol}{numeric, retention time window for EIC}
}
\value{
matrix, retention time column and intensity column
}
\description{
Title: extract EIC
}
