% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractEIC.R
\name{eicRawlcms}
\alias{eicRawlcms}
\title{Title: prepare raw data for EIC}
\usage{
eicRawlcms(mzMLdirectory, mzMLfile)
}
\arguments{
\item{mzMLdirectory}{string, directory of the raw LCMS data}

\item{mzMLfile}{string, file name}
}
\value{
eicRawlcms objected extracted by xcmsRaw
}
\description{
Title: prepare raw data for EIC
}
