% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPeak.R
\name{extractPeak}
\alias{extractPeak}
\title{Title: extract chromatographic peaks with MS2 spectra}
\usage{
extractPeak(
  mzMLdirectory,
  mzMLfile,
  SN = 20,
  ppm = 20,
  peakWidth = c(10, 120),
  mzdiff = 0.005,
  noise = 2000,
  rt_min = 300,
  rt_max = 3000,
  rt_tol = 10
)
}
\arguments{
\item{mzMLdirectory}{string, directory of the raw LCMS data}

\item{mzMLfile}{string, file name}

\item{SN}{numeric, threshold of peak signal-to-noise. default: 20}

\item{ppm}{numeric, xcms maximum difference in m/z for peaks with overlapping retention time. default: 20}

\item{peakWidth}{vector, default: 10 ~ 200}

\item{mzdiff}{numeric, default: 0.005}

\item{noise}{numeric,  noise level, peaks with intensity smaller than noise are omitted. default: 2000}

\item{rt_min}{numeric, lower limit of retention time of interest. default: 300 seconds}

\item{rt_max}{numeric, upper limit of retention time of interest. default: 3000 seconds}

\item{rt_tol}{numeric, retention time window in the assignment of tandem MS. default: 10 seconds}
}
\value{
peaks with MS2 spectra
}
\description{
Title: extract chromatographic peaks with MS2 spectra
}
