% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillGap.R
\name{fillGap}
\alias{fillGap}
\title{Title}
\usage{
fillGap(file_dir, mzmlfiles_pattern, aligned_tb, int_threshold = 0)
}
\arguments{
\item{file_dir}{string, the directory storing the mzML files}

\item{mzmlfiles_pattern}{string, file format of the raw lcms data}

\item{aligned_tb}{data frame, aligned feature table}

\item{int_threshold}{numeric, intensity threshold}
}
\value{
data frame
}
\description{
Title
}
