% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanFeature.R
\name{findAdduct}
\alias{findAdduct}
\title{Title: identify salt adduct ##}
\usage{
findAdduct(chlorine_tb, rawlcms, samNum = 1, rt_tol = 10, mz_tol = 0.01)
}
\arguments{
\item{chlorine_tb}{data frame, Cl peaks summarized in a table}

\item{rawlcms}{xcms object processed by xcmsRaw}

\item{samNum}{numeric}

\item{rt_tol}{numeric, retention time in seconds}

\item{mz_tol}{numeric, m/z  in Da}
}
\value{
data frame with adduct labelled
}
\description{
Title: identify salt adduct ##
}
