% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanFeature.R
\name{findInsourcefrag}
\alias{findInsourcefrag}
\title{Title: find in source fragments}
\usage{
findInsourcefrag(
  peaks,
  chlorine_tb,
  rawfile_dir,
  lcmsfile,
  sampleNumber = 1,
  ISFtable
)
}
\arguments{
\item{peaks}{data frame, all peaks}

\item{chlorine_tb}{data frame, all Cl-containing peaks}

\item{rawfile_dir}{string, file folder storing the raw mzML/mzXML files}

\item{lcmsfile}{string, mzML/mzXML file}

\item{sampleNumber}{numeric, sample number}

\item{ISFtable}{data frame, in source fragment from all peaks}
}
\value{
data frame with in source fragment labelled
}
\description{
Title: find in source fragments
}
