% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanFeature.R
\name{findIsotope}
\alias{findIsotope}
\title{Title: identify isotopes #}
\usage{
findIsotope(chlorine_tb, rawlcms, samNum = 1, rt_tol = 10, mz_tol = 0.01)
}
\arguments{
\item{chlorine_tb}{data frame, Cl peaks summarized in a table}

\item{rawlcms}{xcms object processed by xcmsRaw}

\item{samNum}{numeric}

\item{rt_tol}{numeric, retention time in seconds}

\item{mz_tol}{numeric, m/z  in Da}
}
\value{
data frame with isotope labelled
}
\description{
Title: identify isotopes #
}
