% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNetwork.R
\name{GNPS_score}
\alias{GNPS_score}
\title{Title: GNPS spectral similarity function #####
 library: clue
 input: x,y: data frame containing m/z and intensity in ms2 spectra;}
\usage{
GNPS_score(x, x.pre, y, y.pre, mz.tol = 0.02)
}
\arguments{
\item{x}{data frame, ms2 spectra with m/z column and intensity column}

\item{x.pre}{numeric, precursor ion m/z}

\item{y}{data frame, ms2 spectra with m/z column and intensity column}

\item{y.pre}{numeric, precursor ion m/z}

\item{mz.tol}{numeric, m/z tolerance}
}
\value{
list, score and matched number
}
\description{
Title: GNPS spectral similarity function #####
 library: clue
 input: x,y: data frame containing m/z and intensity in ms2 spectra;
}
