% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunction.R
\name{loadRawMS1}
\alias{loadRawMS1}
\title{Title: load MS1 information}
\usage{
loadRawMS1(mzMLdirectory, mzMLfile)
}
\arguments{
\item{mzMLdirectory}{string, file directory}

\item{mzMLfile}{string, file full name .mzML}
}
\value{
list
}
\description{
Title: load MS1 information
}
