% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractEIC.R
\name{peak_smooth}
\alias{peak_smooth}
\title{Title: smooth the chromatographic peaks}
\usage{
peak_smooth(x, level = 2)
}
\arguments{
\item{x}{vector, a list of intensity information of the peak}

\item{level}{numeric, default 2}
}
\value{
vector, a list of intensity
}
\description{
Title: smooth the chromatographic peaks
}
