% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCl.R
\name{predictCl}
\alias{predictCl}
\title{Title: extract and predict individual peak}
\usage{
predictCl(
  featureInfo,
  rawMS1,
  ms1_rt_v,
  rawlcms,
  ms1_graph_dr,
  binary_model,
  multi_model,
  ms1_spetra_rt_tol = 20,
  ms1_spectra_mass_tol = 25,
  iso_mass_diff_4 = 3.994,
  iso_mass_diff_3 = 3,
  iso_mass_diff_2 = 1.99705,
  iso_mass_diff_1 = 1.003355
)
}
\arguments{
\item{featureInfo}{string,"feature ID", "m/z" and "RT" separated by ";"}

\item{rawMS1}{MS1 spectra object extracted using MSnbase}

\item{ms1_rt_v}{list, retention time list}

\item{rawlcms}{raw LCMS extracted by xcmsRaw function}

\item{ms1_graph_dr}{folder to save MS1 and EIC}

\item{binary_model}{pre-trained binary model to predict the existence of the Cl}

\item{multi_model}{pre-trained multiclass model to predict the number of Cl}

\item{ms1_spetra_rt_tol}{numeric, second unit, retention time window to extract the ms1 spectra}

\item{ms1_spectra_mass_tol}{numeric, Da unit, m/z tolerance to extract the ms1 spectra}

\item{iso_mass_diff_4}{numeric, M4 - M0 m/z difference}

\item{iso_mass_diff_3}{numeric, M3 - M0 m/z difference}

\item{iso_mass_diff_2}{numeric, M2 - M0 m/z difference}

\item{iso_mass_diff_1}{numeric, M1 - M0 m/z difference}
}
\value{
prediction result and MS1 m/z column and RT column
}
\description{
Title: extract and predict individual peak
}
