% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaction_pathways.R
\docType{data}
\name{reaction_pathways}
\alias{reaction_pathways}
\title{51 reaction pathways to create the reaction pathway network}
\format{
A data frame with 51 row and 5 variables
\describe{
     \item{name}{e.g. C6H5}
     \item{formula_change}{e.g. -C6H5+H }
     \item{mass_change}{-76.0313}
     \item{reaction_pathway}{ loss of phenyl }
     \item{category}{reaction or biotransformation}
     }
}
\source{
{Created in-house to serve as an example}
}
\usage{
reaction_pathways
}
\description{
Contains the reaction pathways
}
\examples{
data(reaction_pathways) # Lazy loading. Data becomes visible as soon as
}
\keyword{datasets}
