% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunction.R
\name{rtMS1scan}
\alias{rtMS1scan}
\title{Title: record the retention time of MS1 scan}
\usage{
rtMS1scan(rawMS1)
}
\arguments{
\item{rawMS1}{MSnbase object, list of MS1 spectra}
}
\value{
vector, retention time of all ms1 spectra
}
\description{
Title: record the retention time of MS1 scan
}
