% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectCl.R
\name{selectCl}
\alias{selectCl}
\title{Title: Identify the Cl peaks in a feature table}
\usage{
selectCl(
  mzMLdirectory,
  mzMLfile,
  original_ft,
  binary_model,
  multi_model,
  ms1_spetra_rt_tol = 20,
  ms1_spectra_mass_tol = 25,
  iso_mass_diff_4 = 3.994,
  iso_mass_diff_3 = 3,
  iso_mass_diff_2 = 1.99705,
  iso_mass_diff_1 = 1.003355
)
}
\arguments{
\item{mzMLdirectory}{string, the directory containing the raw mzML/mzXML file}

\item{mzMLfile}{string, name of the raw mzML/mzXML file}

\item{original_ft}{data frame, feature table}

\item{binary_model}{pre-trained binary model to predict the existence of the Cl}

\item{multi_model}{pre-trained multiclass model to predict the number of Cl}

\item{ms1_spetra_rt_tol}{numeric, second unit, retention time window to extract the ms1 spectra}

\item{ms1_spectra_mass_tol}{numeric, Da unit, m/z tolerance to extract the ms1 spectra}

\item{iso_mass_diff_4}{numeric, M4 - M0 m/z difference}

\item{iso_mass_diff_3}{numeric, M3 - M0 m/z difference}

\item{iso_mass_diff_2}{numeric, M2 - M0 m/z difference}

\item{iso_mass_diff_1}{numeric, M1 - M0 m/z difference}
}
\value{
data frame with Cl information
}
\description{
Title: Identify the Cl peaks in a feature table
}
